Exercises 4.15-4.23

+++ part4/blog/backend/controllers/blogs.js
@@ -1,8 +1,15 @@
+const jwt = require('jsonwebtoken')
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog
+    .find({})
+    .populate('user', {
+        username: 1,
+        name: 1
+    })
 
   response.json(blogs)
 })
@@ -14,11 +21,16 @@ blogsRouter.post('/', async (request, response) => {
     return response.status(400).json({ error: 'title or url missing' })
   }
 
+  if (!request.user) {
+    return response.status(401).json({ error: 'invalid invalid' })
+  }
+
   const blog = new Blog({
     title: body.title,
     author: body.author,
     url: body.url,
     likes: body.likes || 0,
+    user: request.user.id
   })
 
   const result = await blog.save()
@@ -26,6 +38,16 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const blog = await Blog.findById(request.params.id)
+
+  if (!blog) {
+    return response.status(404).json({ error: 'blog not found' })
+  }
+
+  if (!request.user || request.user.id !== blog.user.toString()) {
+    return response.status(401).json({ error: 'invalid token' })
+  }
+
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
@@ -43,9 +65,14 @@ blogsRouter.put('/:id', async (request, response) => {
   blog.author = body.author || blog.author
   blog.url = body.url || blog.url
   blog.likes = body.likes || blog.likes
+  blog.user = body.user || blog.user
 
   const result = await blog.save()
-  response.json(result)
+  const updatedBlog = await result.populate('user', {
+    username: 1,
+    name: 1
+  })
+  response.json(updatedBlog)
 })
 
 module.exports = blogsRouter
